+++ appserver/tests/payara-samples/samples/clustered-singleton/clustered-singleton-test/pom.xml
@@ -43,7 +43,6 @@
         <dependency>
             <groupId>org.hamcrest</groupId>
             <artifactId>hamcrest-library</artifactId>
-            <version>1.3</version>
             <scope>test</scope>
         </dependency>
         <dependency>
